Use PATH to figure pyrus.phar's location.
As pyrus.phar may not have been created with execution permissions,
it is not safe to call it directly. Instead, php should be called
with the full path to pyrus.phar.
And, since pyrus.phar probably won't be in the current directly,
the PATH environment variable must be used to figure its location.