Do not ask for passphrase directly.
To make it possible (and easier) to pass a passphrase to Pyrus when using
the "release" target, a new property (named "passfile") has been added.
It should point to the passphrase for the certificate currently set in PEAR.

Should fix an odd issue when generating both a tgz & phar package in turn,
where I suspect the first incovation of pyrus.phar resulted in stdin's
file desc. being closed, making it unusable for the subsequent call.