Auto-add a LICENSE to the final package.
Adds the extrafiles.php script which can be used to add a local LICENSE
file (from the module's point of view) to the soon-to-be package.
The LICENSE file will be added as part of the documentation.

This is mostly useful to ensure the license is distributed along the
package while it sits at the root of the module (so that it stays in
plain sight for all to see), as opposed to being in the docs/ folder
for example.